requirements.txt
flask
openai
azure-search-documents
gunicorn